+++ authorizer.go
@@ -3,8 +3,6 @@ package main
 import (
 	"bytes"
 	"context"
-	"crypto/rand"
-	"encoding/base64"
 	"errors"
 	"io"
 	"net/http"
@@ -43,33 +41,6 @@ func getRequestUserSub(r *http.Request) (*requestUser, error) {
 		return nil, errors.New("user not found")
 	}
 	return r.Context().Value(userKey).(*requestUser), nil
-}
-
-// signinHandler: OIDC proxy backend signin process.
-func signinHandler(w http.ResponseWriter, r *http.Request) {
-	ctx := r.Context()
-	signinUser, err := getRequestUser(r)
-	if err != nil {
-		appLogger.Infof(ctx, "Unauthenticated: %+v", err)
-		http.Error(w, "Unauthenticated", http.StatusUnauthorized)
-		return
-	}
-	token := make([]byte, 24)
-	_, _ = rand.Read(token)
-	http.SetCookie(w, &http.Cookie{
-		Name:   "XSRF-TOKEN",
-		Value:  base64.RawURLEncoding.EncodeToString(token),
-		Path:   "/",
-		Secure: r.Header.Get("X-Forwarded-Proto") == "https",
-	})
-	resp := map[string]interface{}{}
-	if signinUser.userID != 0 {
-		resp["user_id"] = signinUser.userID
-	} else if signinUser.accessTokenID != 0 && signinUser.accessTokenProjectID != 0 {
-		resp["access_token_id"] = signinUser.accessTokenID
-		resp["project_id"] = signinUser.accessTokenProjectID
-	}
-	writeResponse(ctx, w, http.StatusOK, resp)
 }
 
 func (g *gatewayService) UpdateUserFromIdp(next http.Handler) http.Handler {
